Separates reused Bell and Notch behavior assertions into custom Cypress commands
