+++ examples/components/Composition/rendering/behaviors/Notch.test.js
@@ -1,22 +1,8 @@
-export const assertNotch = () => {
-  const assertAllVisible = () => {
-    cy.get('#notch').assertAreas([['left', 'center', 'right']])
-  }
-
-  const assertCenterHidden = () => {
-    cy.get('#notch').assertAreas([['left', 'right']])
-  }
-
-  assertAllVisible()
-  cy.setBreakpoint('sm').then(assertAllVisible)
-  cy.setBreakpoint('md').then(assertCenterHidden)
-  cy.setBreakpoint('lg').then(assertAllVisible)
-  cy.setBreakpoint('xl').then(assertAllVisible)
-}
-
 it('Notch', () => {
   cy.loadStory(
     ['components', 'composition', 'rendering', 'behaviors'],
     ['notch'],
   )
+
+  cy.get('#notch').assertNotch()
 })
