+++ examples/configuration/CustomUnit.test.js
@@ -1,7 +1,5 @@
-import { assertNotch } from '../components/Composition/rendering/behaviors/Notch.test'
-
 it('Supports custom measurement unit', () => {
   cy.loadStory(['configuration', 'custom-configuration'], ['custom-unit'])
 
-  assertNotch()
+  cy.get('#notch').assertNotch()
 })
